my_site/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-mptt==0.13.4
django-recaptcha3==0.4.0
idna==3.3
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
